requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pandas==2.1.3
numpy==1.26.2
matplotlib==3.8.2
seaborn==0.13.0
requests==2.31.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
pydantic==2.5.2
plotly==5.18.0
scikit-learn==1.3.2
scipy==1.11.4
openpyxl==3.1.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
aiofiles==23.2.1
tenacity==8.2.3
colorama==0.4.6
tqdm==4.66.1 